Fixed the warnings in xl_animdefs too

+++ source/xl_animdefs.cpp
@@ -132,7 +132,7 @@ bool XLAnimDefsParser::doStateExpectItem(XLTokenizer &token)
    {
       XLAnimDef def;
       def.type = xlanim_flat;
-      def.index = pics.getLength();
+      def.index = static_cast<int>(pics.getLength());
       defs.add(def);
       curswitch = nullptr;
       curdef = &defs[defs.getLength() - 1];
@@ -145,7 +145,7 @@ bool XLAnimDefsParser::doStateExpectItem(XLTokenizer &token)
    {
       XLAnimDef def;
       def.type = xlanim_texture;
-      def.index = pics.getLength();
+      def.index = static_cast<int>(pics.getLength());
       defs.add(def);
       curswitch = nullptr;
       curdef = &defs[defs.getLength() - 1];
